tests/wpt/tests/html/canvas/offscreen/fill-and-stroke-styles/2d.fillStyle.relativecolor.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.fillStyle.relativecolor</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.fillStyle.relativecolor</h1>
<p class="desc">Relative color works as color input</p>


<script>
var t = async_test("Relative color works as color input");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.fillStyle = "rgb(from red g r b)";
  _assertSame(ctx.fillStyle, 'color(srgb 0 1 0)', "ctx.fillStyle", "'color(srgb 0 1 0)'");
  ctx.fillStyle = "color(from color(srgb 0.25 0.5 0.75 / 0.5) srgb r g b / alpha)";
  _assertSame(ctx.fillStyle, 'color(srgb 0.25 0.5 0.75 / 0.5)', "ctx.fillStyle", "'color(srgb 0.25 0.5 0.75 / 0.5)'");
  t.done();

});
</script>
